# backend/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Team tuner API documentation",
        "title": "TT API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/course": {
            "get": {
                "description": "Get courses",
                "tags": [
                    "courses"
                ],
                "summary": "Get courses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/out.CourseOut"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create course",
                "tags": [
                    "courses"
                ],
                "summary": "Create course",
                "parameters": [
                    {
                        "description": "Create course body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/in.CreateCourseBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/out.CreatedCourseOut"
                        }
                    }
                }
            }
        },
        "/course/{courseId}/group-runs": {
            "get": {
                "description": "Get courses group runs",
                "tags": [
                    "courses"
                ],
                "summary": "Get courses group runs",
                "operationId": "get-courses-group-runs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/out.GroupRunOut"
                            }
                        }
                    }
                }
            }
        },
        "/course/{courseId}/stats/bf-box-plot": {
            "get": {
                "description": "Get a big five box plot for a course",
                "tags": [
                    "courses"
                ],
                "summary": "Get a big five box plot for a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/out.BigFiveBoxPlot"
                        }
                    }
                }
            }
        },
        "/course/{courseId}/stats/daily-activity": {
            "get": {
                "description": "Get a daily number of solved questioners for a course",
                "tags": [
                    "courses"
                ],
                "summary": "Get a daily number of solved questioners for a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "startTime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/out.DailyActivity"
                        }
                    }
                }
            }
        },
        "/course/{courseId}/stats/questioner-stats": {
            "get": {
                "description": "Get a questioner completion stats for a course",
                "tags": [
                    "courses"
                ],
                "summary": "Get a questioner completion stats for a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/out.QuestionerCompletionStats"
                        }
                    }
                }
            }
        },
...
        "/login/github": {
            "post": {
                "description": "Github login",
                "tags": [
                    "login"
                ],
                "summary": "Github login",
                "parameters": [
                    {
                        "description": "Github login body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/in.GithubLoginIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/out.GithubLoginOut"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "in.CreateCourseBody": {
            "type": "object",
            "required": [
                "code",
                "colour",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 10,
                    "minLength": 1
                },
                "colour": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "in.GithubLoginIn": {
            "type": "object",
            "required": [
                "navigateTo"
            ],
            "properties": {
                "navigateTo": {
                    "type": "string"
                }
            }
        },
...
        "out.BigFiveBoxPlot": {
            "type": "object",
            "required": [
                "agreeableness",
                "conscientious",
                "extraversion",
                "neuroticism",
                "openness"
            ],
            "properties": {
                "agreeableness": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                },
                "conscientious": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                },
                "extraversion": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                },
                "neuroticism": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                },
                "openness": {
                    "$ref": "#/definitions/out.BoxPlotItem"
                }
            }
        },
        "out.BoxPlotItem": {
            "type": "object",
            "required": [
                "dataPoints",
                "max",
                "mean",
                "min",
                "q1",
                "q3"
            ],
            "properties": {
                "dataPoints": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "max": {
                    "type": "number"
                },
                "mean": {
                    "type": "number"
                },
                "min": {
                    "type": "number"
                },
                "q1": {
                    "type": "number"
                },
                "q3": {
                    "type": "number"
                }
            }
        },
        "out.CourseOut": {
            "type": "object",
            "required": [
                "code",
                "colour",
                "createdAt",
                "id",
                "name",
                "updatedAt"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "colour": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "out.CreatedCourseOut": {
            "type": "object",
            "required": [
                "code",
                "colour",
                "createdAt",
                "id",
                "name",
                "updatedAt"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "colour": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "out.DailyActivity": {
            "type": "object",
            "required": [
                "activity",
                "endTime",
                "startTime"
            ],
            "properties": {
                "activity": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/out.DailyStat"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "out.DailyStat": {
            "type": "object",
            "required": [
                "count",
                "day"
            ],
            "properties": {
                "count": {
                    "type": "integer"
                },
                "day": {
                    "type": "string"
                }
            }
        },
...
        "out.GithubLoginOut": {
            "type": "object",
            "required": [
                "redirectUrl"
            ],
            "properties": {
                "redirectUrl": {
                    "type": "string"
                }
            }
        },
...
        "out.GroupRunOut": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "updatedAt"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "out.QuestionerCompletionStats": {
            "type": "object",
            "required": [
                "completed",
                "percent",
                "totalCount"
            ],
            "properties": {
                "completed": {
                    "type": "integer"
                },
                "percent": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
    }
}